requirements.txt
streamlit==1.28.1
google-cloud-aiplatform==1.36.2
langchain==0.0.335
tiktoken==0.5.1
openai==1.2.4
google-cloud-language